+++ lib/cogment/services/trial_lifecycle_service.cpp
@@ -22,8 +22,7 @@ namespace cogment {
 TrialLifecycleService::TrialLifecycleService(Orchestrator* orch) : orchestrator_(orch) {}
 
 ::easy_grpc::Future<::cogment::TrialStartReply> TrialLifecycleService::StartTrial(::cogment::TrialStartRequest req,
-                                                                                  easy_grpc::Context ctx) {
-  (void)ctx;
+                                                                                  easy_grpc::Context) {
   auto params = orchestrator_->default_trial_params();
 
   // Apply config override if provided
@@ -48,20 +47,18 @@ TrialLifecycleService::TrialLifecycleService(Orchestrator* orch) : orchestrator_
 
 ::cogment::TerminateTrialReply TrialLifecycleService::TerminateTrial(::cogment::TerminateTrialRequest,
                                                                      easy_grpc::Context ctx) {
-  (void)ctx;
   orchestrator_->end_trial(uuids::uuid::from_string(ctx.get_client_header("trial-id")));
   return {};
 }
 
 ::cogment::TrialInfoReply TrialLifecycleService::GetTrialInfo(::cogment::TrialInfoRequest, easy_grpc::Context ctx) {
-  (void)ctx;
   ::cogment::TrialInfoReply result;
   auto add_trial = [&](Trial* trial) {
     auto trial_info = result.add_trial();
 
     auto trial_lock = trial->lock();
     trial_info->set_trial_id(to_string(trial->id()));
-    trial_info->set_state(get_trial_state_proto(trial->state()));
+    trial_info->set_state(get_trial_api_state(trial->state()));
   };
 
   auto trial_id_str = ctx.get_client_header("trial-id");
@@ -81,14 +78,51 @@ TrialLifecycleService::TrialLifecycleService(Orchestrator* orch) : orchestrator_
   return result;
 }
 
-::easy_grpc::Stream_future<::cogment::TrialListEntry> TrialLifecycleService::WatchTrials(::cogment::TrialListRequest,
-                                                                                         easy_grpc::Context) {
-  throw std::runtime_error("not implemented");
-  return {};
+::easy_grpc::Stream_future<::cogment::TrialListEntry> TrialLifecycleService::WatchTrials(
+    ::cogment::TrialListRequest req, easy_grpc::Context) {
+  // Shared and not a unique ptr because we want the lambda copy constructible
+  auto promise = std::make_shared<Trial_promise>();
+  Trial_future future(promise->get_future());
+
+  // Build a bitmask for testing wether or not a trial should be reported.
+  std::bitset<cogment::TrialState_MAX + 1> state_mask;
+  if (req.filter_size() == 0) {
+    // If filter is empty, we report everything
+    state_mask.set();
+  }
+  else {
+    for (auto state : req.filter()) {
+      state_mask.set(static_cast<std::size_t>(state));
+    }
+  }
+
+  // This will get invoked on each reported trial
+  auto handler = [state_mask, promise](const Trial& trial) {
+    auto state = get_trial_api_state(trial.state());
+
+    if (state_mask.test(static_cast<std::size_t>(state))) {
+      cogment::TrialListEntry msg;
+      msg.set_trial_id(to_string(trial.id()));
+      msg.set_state(state);
+      promise->push(std::move(msg));
+    }
+  };
+
+  // Report current trials.
+  for (const auto& trial : orchestrator_->all_trials()) {
+    handler(*trial.get());
+  }
+
+  // TODO: Between the initial report just above and the handler registration just bellow some trials could change
+  // state. Those states won't be reported
+
+  // Request that the orchestrator invokes the handler whenever trials change.
+  orchestrator_->watch_trials(std::move(handler));
+
+  return future;
 }
 
-::cogment::VersionInfo TrialLifecycleService::Version(::cogment::VersionRequest, easy_grpc::Context ctx) {
-  (void)ctx;
+::cogment::VersionInfo TrialLifecycleService::Version(::cogment::VersionRequest, easy_grpc::Context) {
   ::cogment::VersionInfo result;
   auto v = result.add_versions();