+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+- Watch trials is now supported
+
 ## v1.0.0-alpha5 - 2021-02-19
 
 ### Added